+++ module/core/database/migration/migrations/sqlite/000031_shipping_quotation_remove_discount.up.sql
@@ -1,0 +1,3 @@
+ALTER TABLE shipping_quotations DROP COLUMN full_freight_cost;
+ALTER TABLE shipping_quotations DROP COLUMN discount_percent;
+ALTER TABLE shipping_quotations DROP COLUMN discounted_freight_cost;